rename clonable helper to proxiable
