+++ packages/arbor-store/src/proxiable.ts
@@ -1,0 +1,7 @@
+export default function proxiable(value: any): boolean {
+  return (
+    typeof value === "object" &&
+    value?.constructor != null &&
+    value.constructor.name !== "Date"
+  )
+}
